requirements.txt
opencv-python==4.12.0.88
streamlit==1.41.1
transformers==4.48.0
torch==2.5.1
torchvision==0.20.1
pillow==11.1.0
numpy==2.2.6
pandas==2.2.3
matplotlib==3.10.0
seaborn==0.13.2
ultralytics==8.3.60
requests==2.32.3
